+++ packages/frontend/src/components/table/NoInfoCell.tsx
