+++ packages/frontend/src/pages/scaling-risk/view/ScalingRiskView.tsx
@@ -1,10 +1,10 @@
 import { Layer2, Layer2RiskView } from '@l2beat/config'
 import React from 'react'
 
-import { ProjectLink } from '../../../components/ProjectLink'
-import { RiskCell } from '../../../components/RiskCell'
 import { ScalingLegend } from '../../../components/ScalingLegend'
-import { ColumnConfig, TableView } from '../../../components/TableView'
+import { ProjectLink } from '../../../components/table/ProjectLink'
+import { RiskCell } from '../../../components/table/RiskCell'
+import { ColumnConfig, TableView } from '../../../components/table/TableView'
 
 export interface ScalingRiskViewProps {
   items: ScalingRiskViewEntry[]
